pyproject.toml
[tool.poetry]
name = "chessington"
version = "0.1.0"
description = ""
authors = ["Sam Cappleman-Lynes <sam.cappleman-lynes@softwire.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
start = "chessington.ui:play_game"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
